Make Profile.following symmetrical=False
This fixes a major problem with following. Symmetrical is the Facebook model (you follow me, I follow you), while we want the Twitter model.

Unfortunately, old data will not be fixed with this and it would be impossible to trace all the remote followers to fix local follower counts at this stage.