Fix single content view reply loading issue due to long guids
Earlier fix was made to fix issue with guids that have @ in them. The fix wasn't really done properly and fixing it properly exposed another problem. 'asgiref' library limits Channels group name length to 99 chars.

Do a Content.id + Content.guid to get some extra uniqueness to the group name, then slugify it and cut it at 80 chars (since we prefix it with 'streams_content__'. This should offer enough uniqueness to provide WS notifications to the frontend. Even if a clash happens, only an ID is ever delivered to the frontend, not a whole content.