Restrict logged in user profile search
Filter out LIMITED+SELF profiles if normal logged in user. The rationale, even though those profiles might be visible to view, is that LIMITED profiles are not searchable. We should at some point add those to the list that are in users following list probably.

Make a small change to accessing a profile. LIMITED now is viewable by any logged in user. This makes SITE and LIMITED kind of the same thing and we should maybe remove one of them? LIMITED kind of is unnecessary since we don't recognize sharing as a concept, only following.

Refs: #163