Initial stream precaching
Cache FollowedStream for all active users into redis. This allows very fast output of cached content ID's when loading the stream and enables more heavy calculations to be added what belongs to a stream and what not.

No functionality changes otherwise than moving the views to use the Streams classes instead of querysets directly.

If a Stream class is not cached or there is nothing in the cache, the queryset will be used to pull out enough content id's as before.