Fix regression on visibility of remote replies on shares
Replies inherit the parent object visibility and share visibility defaults to non-public in the federation library. Diaspora protocol removed the ``public`` property from shares in a recent release, which meant that we started getting all shares as non-public from the federation layer. This meant that all comments on the shares were processed as non-public too.

With a change in the federation layer, Diaspora protocol shares are now public by default.