Fix Twitter oembed's in Vue streams
Since the provided async script in oembed's doesn't fire correctly
in Vue streams, we fetch and init the embeds in Stream component
updated method. This is a standard way to embed multiple tweets on
a page from Twitter. For Vue we need some extra tricks and some
binding to make sure imagesLoaded triggers also when the tweet embeds
have finished rendering.

Refs: #202