Add Vue single content view
Make ContentDetailView render the Vue streams app if vue preference
is set. Tweak streams app to render a single content in case the
stream name is content stream. Some small refactoring to optimize
grid stream related actions only when rendering a grid. Small optimization
to store in preparing a 'stream' object when initializing the app.
This will allow to update the state according to routing events
in the future.

Refs: #202