Refactor Twitter OEmbed Masonry layouting
Previous implementation was buggy. Now we load the Twitter widget
script in the base template once, removing need to load it for
each OEmbed. To make sure layouting happens, we do a hacky setTimeout
and do a few layouting calls after 2 and 4 seconds, to ensure
Twitter widget has had time to fetch the tweet.

Refs: #202