Fix Content querysets not correctly including 'through'
This is the information which tells what content caused a share to be added to a stream.

This information was already correctly added in the streams precalculation phase, but if the cache started cold or a viewing user cycled through all cached content ID's and wanted some more, the database queries did not return the right results.

Closes #412