Added betters streams precache maintenance in regards to inactive users
Two new settings have been added:

* ``SOCIALHOME_STREAMS_PRECACHE_INACTIVE_DAYS`` (default 90)
* ``SOCIALHOME_STREAMS_PRECACHE_INACTIVE_SIZE`` (default 0)

If a user has been more than the set days without logging in, when trimming the precaches for that user, the inactive setting will be used instead. By default this means that precaches for users that haven't logged in for 90 days are removed. This is done to ensure Redis memory usage is predictable and stable in relation to active users.

Refs: #436