Refactor tags search
Don't use Haystack index for the tags. It seems Whoosh as a backend
can't handle this amount of items comfortably. Need to think of
separating different types of searches maybe and allowing a different
backend per item, since we absolutely need full content search which
has to be elasticsearch.

For now just use the DB to search for the tags and show the results
separately.