Refactor frontend store
Make the store based on modules, more like a standard Vuex example.
Lots of small changes related to this here and there.

Also place all built resources in a "dist" folder to make it possible
to exclude those in an IDE and to in general make it clearer which
files are sources and which compiled.