Make nested replies support better
Nested replies are now correctly shown as replies to the root level
content instead of being hidden from view. Link all replies at
any level to the root content via `root_parent` foreign key.

In the UI they are flattened and it's not currently possible
to reply to a reply on Socialhome side.