Rewrite local federation ID's
This is a BREAKING CHANGE for anyone who has been running
Socialhome with ActivityPub enabled. Profile federation ID's are
rewritten from `/p/<uuid>` to `/u/<username>`. This is to make the
federation ID's more human understandable.

This of course means that changing username will not be possible and
will break federation. But since many fediverse platforms assume
this anyway, we probably wont be able to add support for that anyway.

Refs: #522