Refactoring of profiles in frontend store
When stream content loads, the new profile store gets populated with
objects. Also load a profile on demand in the
ProfileReactionButtons component.

This refactoring is done to separate more the profiles and the stream
content from each other into separate store modules and to have
less need of bundling all data for non-viewing profiles into the
streams API results.