Switch from `django-debug-toolbar-user-switcher` to `django-debug-toolbar-user-panel`
I had issues getting `django-debug-toolbar-user-switcher` to install. I think we swapped to it due to Django supported versions? The original upstream version has received support up to Django 4 in 2023, so switching back to it probably makes sense given the fork has not been updated for a longer period of time.